+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [Unreleased]
 ### Changed
 - Internal change to move the WebSocket handling into `cable.js`.
+- Build an iframe and change the domain in order to avoid preflight checks.
 
 ## [4.1.0] - 2018-08-11
 ### Added