+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [5.0.1] - 2018-08-13
+### Changed
+- Bind `fetch` properly to the `window` in order to avoid illegal invocation errors.
+- Don't use the `createConsumer` named export from `ActionCable`, just use the main object (doesn't appear to work otherwise).
+
 ## [5.0.0] - 2018-08-13
 ### Added
 - The `skipPreflightChecks` named export that handles building an `iframe` to match domains and skip preflight CORS checks.